apps/raven-api/src/app/core/http-cache.interceptor.ts
import { Observable } from 'rxjs';

import {
  CallHandler,
  ExecutionContext,
  Injectable,
  NestInterceptor,
} from '@nestjs/common';

@Injectable()
export class HttpCacheInterceptor implements NestInterceptor {
  public intercept(
    context: ExecutionContext,
    next: CallHandler
  ): Observable<Response> {
    const request = context.switchToHttp().getRequest();
    const response = context.switchToHttp().getResponse();
    if (request.method === 'GET') {
      response.set('Cache-Control', 'no-cache, no-store, must-revalidate');
    }
    return next.handle();
  }
}
